+++ quint/src/names/collector.ts
@@ -111,6 +111,9 @@ export class NameCollector implements IRVisitor {
 
     const instanceTable = new Map([...moduleTable.entries()])
     if (def.qualifiedName) {
+      // Add the qualifier to `definitionsMyModule` map with a copy of the
+      // definitions, so if there is an export of that qualifier, we know which
+      // definitions to export
       this.definitionsByModule.set(def.qualifiedName, instanceTable)
     }
 
@@ -152,6 +155,14 @@ export class NameCollector implements IRVisitor {
       // Importing non-existing module
       this.errors.push(moduleNotFoundError(def))
       return
+    }
+
+    if (def.qualifiedName) {
+      // Add the qualifier to `definitionsMyModule` map with a copy of the
+      // definitions, so if there is an export of that qualifier, we know which
+      // definitions to export
+      const newTable = new Map([...moduleTable.entries()])
+      this.definitionsByModule.set(def.qualifiedName, newTable)
     }
 
     const qualifier = def.defName ? undefined : def.qualifiedName ?? def.protoName